Moving GenericFactoryExportStrategy into Grace assembly.
